requirements.txt
typing_extensions==4.12.2
ipython==8.26.0
langchain-core==0.3.31
langgraph==0.2.67
langchain==0.3.15
langchain-community==0.3.15
httpx==0.28.0
sounddevice==0.5.1
soundfile==0.12.1
streamlit==1.41.1
langchain-ollama==0.2.2
numpy==1.26.4
pyrubberband==0.4.0
tiktoken==0.8.0
torch==2.1.2
TTS==0.22.0
tqdm==4.66.5
uvicorn==0.30.5
fastapi==0.112.0
python-dotenv==1.0.1
faiss-cpu==1.9.0.post1
langchain-cohere==0.4.2
scikit-learn==1.5.1
pydantic==2.10.6
PyYAML==6.0.1
